chore: fix compose.yml issue with .env files, finalize setup for dev environnement with Docker / Add READMe doc for starting the project with Docker / small changes : rename Dockerfile to Dockerfile.dev, modify api's Dockerfile for gradle build, add dockerignore and gitignore elements, reorganize folders, fix typo, add vite config for docker
