+++ api/.gitignore
@@ -37,4 +37,4 @@ out/
 .vscode/
 
 # dotenv environment variable files
-.env.dev
+.env*
