--- api/Dockerfile
@@ -1,5 +1,0 @@
-FROM openjdk:21-jdk
-WORKDIR /app
-COPY build/libs/*.jar app.jar
-EXPOSE 8080
-ENTRYPOINT ["java", "-jar", "app.jar"]
