+++ api/Dockerfile.dev
@@ -1,0 +1,10 @@
+FROM gradle:8.10.2-jdk-lts-and-current AS build
+WORKDIR /app
+COPY . .
+RUN gradle build --no-daemon
+
+FROM openjdk:21-jdk-slim
+WORKDIR /app
+COPY --from=build /app/build/libs/*.jar app.jar
+EXPOSE 8080
+CMD ["java", "-jar", "app.jar"]
