+++ api/src/main/java/com/marketplace/test/TestController.java
@@ -1,4 +1,4 @@
-package com.marketplace.rest;
+package com.marketplace.test;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
