+++ api/src/main/resources/application.yml
@@ -1,14 +1,14 @@
 spring:
-    application:
-        name: marketplace_backend
+  application:
+    name: marketplace_backend
 
-    datasource:
-        url: jdbc:postgresql://${DB_HOST}:${DB_PORT}/$DB_NAME
-        username: ${DB_USER}
-        password: ${DB_PASSWORD}
-        driver-class-name: org.postgresql.Driver
+  datasource:
+    url: jdbc:postgresql://${DB_HOST}:${DB_PORT}/${DB_NAME}
+    username: ${DB_USER}
+    password: ${DB_PASSWORD}
+    driver-class-name: org.postgresql.Driver
 
-    jpa:
-        hibernate:
-            ddl-auto: update
-        show-sql: true
+  jpa:
+    hibernate:
+      ddl-auto: update
+    show-sql: true
