+++ api/src/test/java/com/marketplace/ApplicationTests.java
@@ -1,13 +1,12 @@
 package com.marketplace;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class ApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+	// @Test
+	// void contextLoads() {
+	// }
 
 }
