+++ client/Dockerfile.dev
@@ -1,9 +1,8 @@
 FROM node:18-alpine
 WORKDIR /app
-COPY package*.json ./
+COPY package.json ./
 RUN npm install
-COPY .. .
-
-CMD ["npm", "run", "start"]
+COPY . .
 EXPOSE 5173
+CMD ["npm", "run", "dev"]
